Rewrite
Rewriting the module to remove all changes to the database and to make it faster to use. Each class will select a specialization to get their equipment based on that specialization. The weapons might not be perfect for that spec, but in that case it's easy to change. It's also meant to be replaced, not permanent. It will not allow you to get your equipment if you have some equipped, both to prevent overwriting better equipment but also to prevent exploits. For some reason, equipping equipment while you still have equipment equipped will stack the stats. That won't be the case with this module. The equipment function is only available to players who are level 80.

I still need to add the code for gems and glyphs, which will take a while.

I also moved all variables into a header to help with readability, I guess.