Implementations
Added Meta to the list of Gems.
When using the functions, the player will be sent back to the previous menu instead of closing the menu every time.
Fixed the removing of equipment. It wasn't saving when removing the equipment so relogging kept bringing the equipment back. Now it should work as intended.